+++ NEWS.md
@@ -1,3 +1,11 @@
+# openmeteo 0.2.0
+
+* Added support for additional APIs:
+  * Marine Weather
+  * Climate Forecasts
+  * Floods
+  * Air Quality
+
 # openmeteo 0.1.0
 
 * First public release.